Reports on a likelihood ratio test of the mixed model vs. a model that does not account for nesting

Formal test of whether the model that accounts for the correlation of responses within the same unit fits the data better than a model that assumes 0 correlation between responses within the same unit.

levelCompare(model)

Arguments

model
A linear mixed-effects model of class lmerMod or lmerModLmerTest

Value

Computes a likelihood ratio test of the requested model compared to a model with the same predictors but absent the random intercept(s) and slope(s). The p-value of this test is reported.

Details

The likelihood ratio test is Chi-square distributed with as many degrees of freedom as there are random variances and covarances estimated. In a two-level model with a single random intercept, the likelihood ratio sampling distribution is Chi-square with 1 degree of freedom. levelCompare uses the anova() function to perform the likelihood ratio test. Note that because a variance cannot be negative, the actual reference distribution is a 50:50 mix of a spike at 0 and a Chi-square with 1 degree of freedom. The p-value reported does not take this into account and should be divided by 2. See Rabe-Hesketh & Skrondal (2012) Chapter 2, section 2.6.2 for further detail.

References

Rabe-Hesketh, S. & Skrondal, A. (2012) Multilevel and Longitudinal Modeling Using Stata, Third Ed., Volume 1. College Station, TX: Stata Press.

Examples

# \donttest{
data(instruction)
mod <- lme4::lmer(mathgain ~ (1 | classid), data = instruction)
levelCompare(mod)
#> refitting model(s) with ML (instead of REML)
#> Chisq(1) = 32.47, p < .001, logLik = -5889.46.
#> The model that accounts for nesting (lmer model) fits the data significantly 
#> better than a model that does not account for nesting (lm model). 
#> This suggests that the random-effects model is needed to account for the 
#> observed nesting structure.
# }